Support multiple-column partitions
This removes any dependnecy on the auto-increment value, as AI can only tell
us a single column's position, and for multi-column partitions we'll need
more than that.

This change does not totally remove the get_autoincrement method, as we'll
still want to confirm the table has the partitioned feature, which we'll do
in a next commit.