Let bootstrap work on tables that aren't partitioned yet.
Add a --assume-partitioned-on flag to bootstrap, to facilitate operation on
tables that do not yet have a partition map. One needs to supply
`--assume-partitioned-on COLUMN_NAME` as many times as needed to identify all
the columns which will be part of the partition expression.

The 'bootstrap' command now emits table-copy instructions

The original plan for 'bootstrap' was to do live alterations, that they should
only lock what they needed, however InnoDB likes to lock everything. So instead
we need to always assume that "bootstrapping" will be a live table clone,
and at their conclusion the team should perform an atomic rename.

Fix unit test to not depend on time of day (oops)

Catch an arithmetic error

Emit output lines with MySQL comment characters